2025.10 deprecations (#3407)
This PR adds information about the APIs that are planned to be removed in oneDAL 2026.0.
All those APIs are now marked as deprecated in the headers and in the docs.

The description of the deprecations added into the documentation:
docs/source/deprecation.rst‎

Also:

-Wno-deprecated-declarations flag was added into Bazel config, and
/wd4996 flag was added to DAAL examples CMake on Windows
in order to suppress the deprecation warnings.

See the respective RFCS for the details:
https://github.com/uxlfoundation/oneDAL/tree/rfcs/rfcs/20251015-DAAL-algorithms-deprecation
https://github.com/uxlfoundation/oneDAL/tree/rfcs/rfcs/20251016-DAAL-data-sources-deprecation
https://github.com/uxlfoundation/oneDAL/tree/rfcs/rfcs/20251023-DAAL-matrices-deprecation
https://github.com/uxlfoundation/oneDAL/tree/rfcs/rfcs/20251022-DAAL-interfaces-deprecation